Create default implementation of load()
